requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.5

# Document Processing
python-docx==1.1.0
pdfplumber==0.11.0

# AI/ML
openai==1.10.0

# Text Processing
markdown==3.5.2
beautifulsoup4==4.14.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-multipart==0.0.6

# Testing
pytest==7.4.4
httpx==0.26.0
